AI/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
distro==1.9.0
environs==14.1.1
executing==2.2.0
fastapi==0.115.8
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
jiter==0.8.2
jupyter_client==8.6.3
jupyter_core==5.7.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
openai==1.63.0
packaging==24.2
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.50
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32==308
pyzmq==26.2.1
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.45.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
